+++ FrameworkFree/Models/Entities/Msg.cs
@@ -4,9 +4,9 @@ namespace Own.Database
     public partial class Msg
     {
         [Key]
-        public int Id { get; set; }
-        public int ThreadId { get; set; }
-        public int AccountId { get; set; }
+        public long Id { get; set; }
+        public long ThreadId { get; set; }
+        public long AccountId { get; set; }
         public string MsgText { get; set; }
     }
 }
